Cleanup Editor files
