+++ AuroraEditor/Base/Editor/CodeEditorView/UI/EditorAccessories/Minimap/MinimapTypeSetter.swift
@@ -1,82 +1,12 @@
 //
-//  MinimapView.swift
-//  
+//  MinimapTypeSetter.swift
+//  AuroraEditor
 //
-//  Created by Manuel M T Chakravarty on 05/05/2021.
+//  Created by Nanashi Li on 2022/09/24.
+//  Copyright © 2022 Aurora Company. All rights reserved.
 //
-//  TextKit subclasses to implement minimap functionality. This is currently only supported for macOS as iOS lacks
-//  support for configuring the type setter.
 
-import SwiftUI
-
-/// Customised text view for the minimap.
-class MinimapView: NSTextView {
-    weak var codeView: CodeView?
-
-    // Highlight the current line.
-    override func drawBackground(in rect: NSRect) {
-        super.drawBackground(in: rect)
-
-        guard let layoutManager = layoutManager else { return }
-
-        // Highlight the current line
-        codeView?.theme.currentLineColour.setFill()
-        if let location = insertionPoint {
-            layoutManager.enumerateFragmentRects(forLineContaining: location) { rect in
-                NSBezierPath(rect: rect).fill()
-            }
-        }
-    }
-}
-
-/// Customised layout manager for the minimap.
-class MinimapLayoutManager: NSLayoutManager {
-
-    // In place of drawing the actual glyphs, we draw small rectangles in the glyph's foreground colour. We ignore the
-    // actual glyph metrics and draw all glyphs as a fixed-sized rectangle whose height is determined by the "used
-    // rectangle" and whose width is a fraction of the actual (monospaced) font of the glyph (rounded to full points).
-    override func drawGlyphs(forGlyphRange glyphsToShow: NSRange, at origin: CGPoint) {
-        guard let textStorage = self.textStorage else { return }
-
-        // Compute the width of a single rectangle representing one character in the original text display.
-        var width: CGFloat
-        let charIndex = self.characterIndexForGlyph(at: glyphsToShow.location)
-        if let font = textStorage.attribute(.font, at: charIndex, effectiveRange: nil) as? NSFont {
-
-            width = minimapFontSize(for: font.pointSize) / 2
-
-        } else { width = 1 }
-
-        globalMainQueue.async {
-            self.enumerateLineFragments(forGlyphRange: glyphsToShow) { (_, usedRect, _, glyphRange, _) in
-
-                let origin = usedRect.origin
-                for index in 0..<glyphRange.length {
-
-                    // We don't draw hiden glyphs (`.null`), control chracters, and "elastic" glyphs, where the latter
-                    // serve as a proxy for white space
-                    let property = self.propertyForGlyph(at: glyphRange.location + index)
-                    if property != .null && property != .controlCharacter && property != .elastic {
-
-                        // TODO: could try to optimise by using the `effectiveRange` of \
-                        // the attribute lookup to compute an entire glyph run to draw as one rectangle
-                        let charIndex = self.characterIndexForGlyph(at: glyphRange.location + index)
-                        if let colour = textStorage.attribute(.foregroundColor,
-                                                              at: charIndex,
-                                                              effectiveRange: nil) as? NSColor {
-                            colour.withAlphaComponent(0.30).setFill()
-                        }
-                        NSBezierPath(rect: CGRect(x: origin.x + CGFloat(index),
-                                                  y: origin.y,
-                                                  width: width,
-                                                  height: usedRect.size.height))
-                        .fill()
-                    }
-                }
-            }
-        }
-    }
-}
+import AppKit
 
 class MinimapTypeSetter: NSATSTypesetter {
 
@@ -276,27 +206,4 @@ class MinimapTypeSetter: NSATSTypesetter {
         lineFragmentRect.pointee.size.height = lineHeight
         lineFragmentUsedRect.pointee.size.height = fontHeight
     }
-}
-
-/// Compute the size of the code view in number of characters such that we can still accommodate the minimap.
-///
-/// - Parameters:
-///   - width: Overall width available for main and minimap code view *without* gutter and padding.
-///   - font: The fixed pitch font of the main text view.
-///   - withMinimap: Determines whether to include the presence of a minimap into the calculation.
-/// - Returns: The width of the code view in number of characters.
-func codeWidthInCharacters(for width: CGFloat, with font: NSFont, withMinimap: Bool) -> CGFloat {
-    let minimapCharWidth = withMinimap ? minimapFontSize(for: font.pointSize) / 2 : 0
-    return floor(width / (font.maximumAdvancement.width + minimapCharWidth))
-}
-
-/// Compute the font size for the minimap from the font size of the main text view.
-///
-/// - Parameter fontSize: The font size of the main text view
-/// - Returns: The font size for the minimap
-///
-/// The result is always divisible by two, to enable the use of full pixels for the font width while avoiding aspect
-/// ratios that are too unbalanced.
-func minimapFontSize(for fontSize: CGFloat) -> CGFloat {
-    return max(1, ceil(fontSize / 20)) * 2
 }
